release版本V0.1.0
1.发布release版本:V0.1.0
2.增加配置功能
3.增加版本号
4.提交协议文档
5.添加发布hex

+++ code/stm32F4xx/stm32f4xx/projects/stm32f4xx/source/fml/transfor/transfor_makepackage.h
@@ -66,7 +66,7 @@ typedef struct
 	float speed_peak;
 	float offset_peak;
 	float temperature;
-	float power;
+	uint8_t  power;
 	uint8_t sum_check;
 	uint8_t foot;
 }Trans485_dataform_t;
@@ -90,7 +90,7 @@ typedef struct
 	float speed_peak;
 	float offset_peak;
 	float temperature;
-	float power;
+	uint8_t  power;
 }Trans485_datavalue_t;
 
 #pragma pack()
